Move OVFUtils.py from core imgfac to ovfcommon pseudo-plugin
