+++ imagefactory_plugins/RHEVM/RHEVMHelper.py
@@ -14,7 +14,7 @@ from ovirtsdk.xml import params
 from xml.etree import ElementTree
 from time import sleep
 from threading import BoundedSemaphore
-from imgfac.OVFUtils import RHEVOVFPackage
+from imagefactory_plugins.ovfcommon.ovfcommon import RHEVOVFPackage
 
 # Large portions derived from dc-rhev-img from iwhd written by
 # Pete Zaitcev <zaitcev@redhat.com>
