+++ imagefactory_plugins/imagefactory-plugins.spec.in
@@ -32,12 +32,23 @@ imagefactory allows the creation of system images for multiple virtualization
 and cloud providers from a single template definition. See 
 http://aeolusproject.org/projects/imagefactory for more information.
 
+%package ovfcommon
+Summary: common utilities to manipulate ovf-related objects
+License: ASL 2.0
+Requires: oz >= 0.7.0
+Requires: imagefactory
+Requires: imagefactory-plugins
+
+%description ovfcommon
+This pseudo-plugin is used to provide common OVF functionality to other plugins.
+
 %package OVA
 Summary: Cloud plugin for generating OVA archives
 License: ASL 2.0
 Requires: oz >= 0.7.0
 Requires: imagefactory
 Requires: imagefactory-plugins
+Requires: imagefactory-plugins-ovfcommon
 
 %description OVA
 This Cloud plugin allows users to specify a Base Image to generate an OVA archive from.
@@ -120,6 +131,7 @@ Summary: RHEVM Cloud plugin
 License: ASL 2.0
 Requires: imagefactory
 Requires: imagefactory-plugins
+Requires: imagefactory-plugins-ovfcommon
 Requires: ovirt-engine-sdk >= 3.1.0
 
 %description RHEVM
@@ -216,6 +228,9 @@ python setup.py install -O1 --root=%{buildroot} --skip-build
 %dir %{python_sitelib}/imagefactory_plugins
 %{python_sitelib}/imagefactory_plugins/__init__.py*
 %{python_sitelib}/imagefactory_plugins*.egg-info
+
+%files ovfcommon
+%{python_sitelib}/imagefactory_plugins/ovfcommon/*
 
 %files OVA
 %{python_sitelib}/imagefactory_plugins/OVA/*
