+++ imagefactory_plugins/ovfcommon/ovfcommon.py
